requirements.txt
requests
Pillow
python-telegram-bot[job-queue]
beautifulsoup4
lxml
python-dotenv
